Task2_SamsungPhoneAdvisor/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
openai
spacy
rapidfuzz
pandas
pytest
requests
beautifulsoup4
lxml
python-multipart
jinja2